Goblin follow mechanic now works
/ Renamed EnemyGoblin scriptable object to EnemyGoblinOld
/ Renamed GoblinController script to GoblinControllerOld

+ Added TestingGoblinAI scene
+ Added new EnemyGoblin scriptable object
+ Added new Goblin Controller
+ Added GoblinFollowRadiusCollider script

- Removed TestingCharacters scene
- Removed SphereGizmo script

Todo:
* Add attack functionality
* Remove EnemyGoblinOld scriptable object
* Remove GoblinControllerOld script
* Remove AIVision script - this has been implemented directly into the new GoblinController script
* Remove unnecessary debug code from new GoblinController script after attack functionality has been added

+++ Goblin Gauntlet/Assets/Scripts/Character.cs
@@ -7,11 +7,11 @@ public abstract class Character : ScriptableObject
 	// Shared characteristics between all characters
 	public string id;
 	public string characterName;
-	public int health;
+	public float health;
 	public float speed;
 	public float rotationSpeed;
 
-	public virtual void TakeDamage(int amount)
+	public virtual void TakeDamage(float amount)
 	{
 		health -= amount;
 	}
